+++ RedBugs_webapp/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+spring.web.resources.static-locations=classpath:/templates/static
