--- RedBugs_webapp/src/main/resources/public/error/404.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-Wrong page, buddy;
-</body>
-</html>
