+++ RedBugs_webapp/src/main/resources/templates/data.html
@@ -13,7 +13,7 @@
 </head>
 <body>
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
-    <a class="navbar-brand" href="/index.html">Redbugs</a>
+    <a class="navbar-brand" href="../index.html">Redbugs</a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarText"
             aria-controls="navbarText" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
@@ -21,13 +21,13 @@
     <div class="collapse navbar-collapse" id="navbarText">
         <ul class="navbar-nav mr-auto">
             <li class="nav-item active">
-                <a class="nav-link" href="/tripFinder.html">Strona główna</a>
+                <a class="nav-link" href="../tripFinder.html">Strona główna</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="/results.html">Rozkład jazdy</a>
+                <a class="nav-link" href="../results.html">Rozkład jazdy</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="data.html">Trasy</a>
+                <a class="nav-link" href="../data.html">Trasy</a>
             </li>
         </ul>
     </div>
