+++ RedBugs_webapp/src/main/resources/templates/error/404.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>404 Page not found</title>
+    <link
+            href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css"
+            rel="stylesheet"
+            integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65"
+            crossorigin="anonymous"
+    />
+</head>
+<body>
+<div class="container">
+    <div class="d-flex justify-content-center">
+        <h1>404 Page not found</h1>
+    </div>
+</div>
+</body>
+</html>
