--- RedBugs_webapp/src/main/resources/templates/redbugsweb.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-
-<button type="button" class="btn btn-primary btn-lg">Login</button>
-<button type="button" class="btn btn-secondary btn-lg">Register</button>
-
-<button type="button" class="btn btn-primary btn-lg">Schedule</button>
-<button type="button" class="btn btn-secondary btn-lg">Bus stop list</button>
-
-<button type="button" class="btn btn-primary btn-lg">Route</button>
-<button type="button" class="btn btn-secondary btn-lg">Exit</button>
-</body>
-</html>
