+++ RedBugs_webapp/src/main/resources/templates/static/index.html
