+++ RedBugs_webapp/src/main/resources/templates/static/results.html
@@ -9,6 +9,7 @@
           integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65"
           crossorigin="anonymous"
   />
+
 </head>
 <body>
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
@@ -26,7 +27,7 @@
         <a class="nav-link" href="results.html">Rozkład jazdy</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="../templates/data.html">Trasy</a>
+        <a class="nav-link" href="templates/data.html">Trasy</a>
       </li>
     </ul>
   </div>
