+++ RedBugs_webapp/src/main/resources/templates/static/tripFinder.html
@@ -26,7 +26,7 @@
         <a class="nav-link" href="results.html">Rozkład jazdy</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="../templates/data.html">Trasy</a>
+        <a class="nav-link" href="templates/data.html">Trasy</a>
       </li>
     </ul>
   </div>
