+++ RedBugs_webapp/src/main/resources/templates/static/user.html
